fixed ffmpeg_concat errors on windows
On windows calling ffmpeg shell commands with os.system using the paths returned by os.path.join doesn't work. Need to change the slashes. 
one way is to use sys.platform to check if it's on windows and then change the slashes as done in the section # windows specific path manipulation